+++ frontend/src/metabase/common/components/DatePicker/DateOperatorPicker/constants.ts
@@ -1,7 +1,7 @@
 import { t } from "ttag";
-import type { Option } from "./types";
+import type { OperatorOption } from "./types";
 
-export const OPTIONS: Option[] = [
+export const OPERATOR_OPTIONS: OperatorOption[] = [
   {
     label: t`All time`,
     value: "none",
