+++ frontend/src/metabase/common/components/DatePicker/DateOperatorPicker/index.ts
@@ -1,0 +1,1 @@
+export * from "./DateOperatorPicker";
