+++ frontend/src/metabase/common/components/DatePicker/DateOperatorPicker/types.ts
@@ -1,0 +1,14 @@
+import type { DatePickerOperator } from "../types";
+
+export type OperatorType =
+  | DatePickerOperator
+  | "none"
+  | "last"
+  | "next"
+  | "current";
+
+export interface OperatorOption {
+  label: string;
+  value: OperatorType;
+  operators: DatePickerOperator[];
+}
