--- frontend/src/metabase/common/components/DatePicker/SimpleDatePicker/index.ts
@@ -1,1 +1,0 @@
-export * from "./SimpleDatePicker";
