+++ frontend/src/metabase/common/components/DatePicker/index.ts
@@ -1,4 +1,3 @@
 export * from "./DatePicker";
-export * from "./SimpleDatePicker";
 export * from "./types";
 export * from "./utils";
